monitor y mensajes

+++ trunk/phpost/themes/state_mega/templates/admin_mods/m.mod_report_posts.php
@@ -9,8 +9,10 @@
                                 <div style="display: block;" class="mensajes ok">Tus cambios han sido guardados.</div>
                                 <?
                                 }
+                                
+                                if ($tsAct == '')
+                                {
                                 ?>
-                                {if $tsAct == ''}
                                     Recuerda leer el protocolo para poder moderar los post que han sido denunciados por otros usuarios, 
                                     si te es posible y se puede editar un post no lo borres, <b>Editalo!</b> 
                                     <hr class="separator" />
@@ -31,27 +33,38 @@
                                             ?>    
                                             <tr id="report_<? echo $r['post_id']; ?>">
                                             	<td><? echo $r['total']; ?></td>
-                                                <td><a href="<? echo $tsConfig['url']; ?>/posts/{$r.c_seo}/{$r.post_id}/{$r.post_title|seo}.html" target="_blank">{$r.post_title|truncate:30}</a></td>
-                                                <td>{$r.d_date|hace:true}</td>
-                                                <td>{$tsDenuncias[$r.d_razon]}</td>
+                                                <td><a href="<? echo $tsConfig['url']; ?>/posts/<? echo $r['c_seo']; ?>/<? echo $r['post_id']; ?>/<? echo string_seo($r['post_title']); ?>.html" target="_blank"><? echo string_truncate($r['post_title'],30); ?></a></td>
+                                                <td><? echo modifier_hace($r['d_date']); ?></td>
+                                                <td><? echo $tsDenuncias[$r['d_razon']]; ?></td>
                                                 <td class="admin_actions">
                                                     <a href="<? echo $tsConfig['url']; ?>/moderacion/posts?act=info&obj=<? echo $r['post_id']; ?>">
                                                         <img src="<? echo $tsConfig['default']; ?>/images/icons/details.png" title="Ver Detalles" /></a>
                                                     <a href="#" onclick="mod.posts.view(<? echo $r['post_id']; ?>); return false;">
                                                         <img src="<? echo $tsConfig['default']; ?>/images/icons/find.png" title="Ver Post" /></a>
-                                                    {if $tsUser->is_admod || $tsUser->permisos.mocdp}
+                                                    <?
+                                                    if ($tsUser->is_admod || $tsUser->permisos['mocdp'])
+                                                    {
+                                                    ?>
                                                     <a href="#" onclick="mod.reboot(<? echo $r['post_id']; ?>, 'posts', 'reboot', false); return false;">
-                                                        <img src="<? echo $tsConfig['default']; ?>/images/icons/reboot.png" title="{if $r.post_status == 1}Reactivar Post{else}Desechar denuncias{/if}" />
+                                                        <img src="<? echo $tsConfig['default']; ?>/images/icons/reboot.png" title="<? if ($r['post_status'] == 1) echo 'Reactivar Post'; else echo 'Desechar denuncias'; ?>" />
                                                     </a>
-                                                    {/if}
-                                                    {if $tsUser->is_admod || $tsUser->permisos.moedpo}
+                                                    <?
+                                                    }
+                                                    if ($tsUser->is_admod || $tsUser->permisos['moedpo'])
+                                                    {
+                                                    ?>
                                                     <a href="<? echo $tsConfig['url']; ?>/posts/editar/<? echo $r['post_id']; ?>" target="_blank">
                                                         <img src="<? echo $tsConfig['default']; ?>/images/icons/edit.png" title="Editar Post" /></a>
-                                                    {/if}
-                                                    {if $tsUser->is_admod || $tsUser->permisos.moep}
+                                                    <?
+                                                    }
+                                                    if ($tsUser->is_admod || $tsUser->permisos['moep'])
+                                                    {
+                                                    ?>
                                                     <a href="#" onclick="mod.posts.borrar(<? echo $r['post_id']; ?>, false); return false">
                                                         <img src="<? echo $tsConfig['default']; ?>/images/icons/close.png" title="Borrar Post" /></a>
-                                                    {/if}
+                                                    <?
+                                                    }
+                                                    ?>
                                                 </td>
                                             </tr>
                                             <?                                            
@@ -71,11 +84,15 @@
                                             <th colspan="5">&nbsp;</th>
                                         </tfoot>
                                     </table>
-                                    {elseif $tsAct == 'info'}
+                                    <?
+                                    }
+                                    elseif ($tsAct == 'info')
+                                    {
+                                    ?>
                                     <h2 style="border-bottom:1px dashed #CCC; padding-bottom:5px;">
-                                        <a href="<? echo $tsConfig['url']; ?>/posts/{$tsDenuncia.data.c_seo}/<? echo $tsDenuncia['data']['post_id']; ?>/{$tsDenuncia.data.post_title|seo}.html" target="_blank">
+                                        <a href="<? echo $tsConfig['url']; ?>/posts/<? echo $tsDenuncia['data']['c_seo']; ?>/<? echo $tsDenuncia['data']['post_id']; ?>/{$tsDenuncia.data.post_title|seo}.html" target="_blank">
                                             <? echo $tsDenuncia['data']['post_title']; ?></a> de 
-                                        <a href="<? echo $tsConfig['url']; ?>/perfil/{$tsDenuncia.data.user_name}">
+                                        <a href="<? echo $tsConfig['url']; ?>/perfil/<? echo $tsDenuncia['data']['user_name']; ?>">
                                             <? echo $tsDenuncia['data']['user_name']; ?>
                                         </a> 
                                         <span class="floatR admin_actions">
@@ -86,14 +103,14 @@
                                             {
                                             ?>
                                             <a href="#" onclick="mod.reboot(<? echo $tsDenuncia['data']['post_id']; ?>, 'posts', 'reboot', true); return false">
-                                                <img src="<? echo $tsConfig['default']; ?>/images/icons/reboot.png" title="{if $tsDenuncia.data.post_status == 1}Reactivar Post{else}Desechar denuncias{/if}" />
+                                                <img src="<? echo $tsConfig['default']; ?>/images/icons/reboot.png" title="<? if ($tsDenuncia['data']['post_status'] == 1) echo 'Reactivar Post'; else echo 'Desechar denuncias';?>" />
                                             </a>
                                             <?
                                             }
                                             if ($tsUser->is_admod || $tsUser->permisos['moedpo'])
                                             {
                                             ?>
-                                            <a href="<? echo $tsConfig['url']; ?>/posts/editar/{$tsDenuncia.data.post_id}" target="_blank">
+                                            <a href="<? echo $tsConfig['url']; ?>/posts/editar/<? echo $tsDenuncia['data']['post_id']; ?>" target="_blank">
                                                 <img src="<? echo $tsConfig['default']; ?>/images/icons/edit.png" title="Editar Post" /></a>
                                             <?                                            
                                             }
@@ -121,10 +138,10 @@
                                                 {
                                                 ?>
                                             <tr>
-                                            	<td><a href="<? echo $tsConfig['url']; ?>/perfil/{$d.user_name}"><? echo $d['user_name']; ?></a></td>
-                                                <td>{$tsDenuncias[$d.d_razon]}</td>
-                                                <td>{$d.d_extra}</td>
-                                                <td>{$d.d_date|hace:true}</td>
+                                            	<td><a href="<? echo $tsConfig['url']; ?>/perfil/<? echo $d['user_name']; ?>"><? echo $d['user_name']; ?></a></td>
+                                                <td><? echo $tsDenuncias[$d['d_razon']]; ?></td>
+                                                <td><? echo $d['d_extra']; ?></td>
+                                                <td><? echo modifier_hace($d['d_date']); ?></td>
                                             </tr>
                                             <?
                                             }
@@ -134,5 +151,7 @@
                                             <th colspan="5">&nbsp;</th>
                                         </tfoot>
                                     </table>
-                                    {/if}
+                                    <?
+                                    }
+                                    ?>
                                 </div>
